+++ scripts/vmware.sh
@@ -1,0 +1,8 @@
+#! /usr/bin/env bash
+
+e() {
+  echo -e "  \e[1;32m[install] ==> $@\e[0m";
+}
+
+e "Install VMware Tools"
+yum install -y open-vm-tools